Delete api/debug/*
