+++ api/controller/trigger.go
@@ -8,6 +8,7 @@ import (
 	"github.com/moira-alert/moira/api"
 	"github.com/moira-alert/moira/api/dto"
 	"github.com/moira-alert/moira/database"
+	"github.com/moira-alert/moira/support"
 )
 
 // UpdateTrigger update trigger data and trigger metrics in last state
@@ -162,4 +163,22 @@ func SetTriggerMaintenance(database moira.Database, triggerID string, triggerMai
 		return api.ErrorInternalServer(err)
 	}
 	return nil
+}
+
+//PullTrigger returns raw trigger from database
+func PullTrigger(database moira.Database, logger moira.Logger, triggerID string) (*moira.Trigger, *api.ErrorResponse) {
+	trigger, err := support.HandlePullTrigger(logger, database, triggerID)
+	if err != nil {
+		return nil, api.ErrorInternalServer(err)
+	}
+	return trigger, nil
+}
+
+//PullTrigger returns raw trigger metrics from database
+func PullTriggerMetrics(database moira.Database, logger moira.Logger, triggerID string) ([]support.PatternMetrics, *api.ErrorResponse) {
+	metrics, err := support.HandlePullTriggerMetrics(logger, database, triggerID)
+	if err != nil {
+		return nil, api.ErrorInternalServer(err)
+	}
+	return metrics, nil
 }
